+++ dev/src/SignalBinding.js
@@ -62,6 +62,12 @@
          * @type boolean
          */
         active : true,
+        
+        /**
+         * Default parameters passed to listener during `Signal.dispatch` and `SignalBinding.execute`. (curried parameters)
+         * @type Array|null
+         */
+        params : null,
 
         /**
          * Call listener passing arbitrary parameters.
@@ -70,14 +76,15 @@
          * @return {*} Value returned by the listener.
          */
         execute : function (paramsArr) {
-            var r;
+            var handlerReturn, params;
             if (this.active && !!this._listener) {
-                r = this._listener.apply(this.context, paramsArr);
+                params = this.params? this.params.concat(paramsArr) : paramsArr;
+                handlerReturn = this._listener.apply(this.context, params);
                 if (this._isOnce) {
                     this.detach();
                 }
             }
-            return r;
+            return handlerReturn;
         },
 
         /**